Madlibs bug fixes
Fixed the first do-while loop in the getMadlibsPoem function so
that when getting the first and last lines of the poem, it continues
looping while the selected line is either blank OR one-character long.
Previously this only looped while both conditions were true.

Also stopped the typeText function when the user goes into Madlibs mode.